+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CleverTapInstanceConfig.java
@@ -49,6 +49,8 @@ public class CleverTapInstanceConfig implements Parcelable {
 
     private String spikyProxyDomain;
 
+    private String customHandshakeDomain;
+
     @NonNull
     private ArrayList<String> allowedPushTypes = getAll();
 
@@ -113,6 +115,7 @@ public class CleverTapInstanceConfig implements Parcelable {
         this.accountRegion = config.accountRegion;
         this.proxyDomain = config.proxyDomain;
         this.spikyProxyDomain = config.spikyProxyDomain;
+        this.customHandshakeDomain = config.customHandshakeDomain;
         this.isDefaultInstance = config.isDefaultInstance;
         this.analyticsOnly = config.analyticsOnly;
         this.personalization = config.personalization;
@@ -181,6 +184,9 @@ public class CleverTapInstanceConfig implements Parcelable {
             if (configJsonObject.has(Constants.KEY_SPIKY_PROXY_DOMAIN)) {
                 this.spikyProxyDomain = configJsonObject.getString(Constants.KEY_SPIKY_PROXY_DOMAIN);
             }
+            if (configJsonObject.has(Constants.KEY_CUSTOM_HANDSHAKE_DOMAIN)) {
+                this.customHandshakeDomain = configJsonObject.optString(Constants.KEY_CUSTOM_HANDSHAKE_DOMAIN, null);
+            }
             if (configJsonObject.has(Constants.KEY_ACCOUNT_REGION)) {
                 this.accountRegion = configJsonObject.getString(Constants.KEY_ACCOUNT_REGION);
             }
@@ -243,6 +249,7 @@ public class CleverTapInstanceConfig implements Parcelable {
         accountRegion = in.readString();
         proxyDomain = in.readString();
         spikyProxyDomain = in.readString();
+        customHandshakeDomain = in.readString();
         analyticsOnly = in.readByte() != 0x00;
         isDefaultInstance = in.readByte() != 0x00;
         useGoogleAdId = in.readByte() != 0x00;
@@ -311,6 +318,14 @@ public class CleverTapInstanceConfig implements Parcelable {
 
     public void setSpikyProxyDomain(String spikyProxyDomain) {
         this.spikyProxyDomain = spikyProxyDomain;
+    }
+
+    public String getCustomHandshakeDomain() {
+        return customHandshakeDomain;
+    }
+
+    public void setCustomHandshakeDomain(String handshakeDomain) {
+        this.customHandshakeDomain = handshakeDomain;
     }
 
     @SuppressWarnings({"unused"})
@@ -394,6 +409,7 @@ public class CleverTapInstanceConfig implements Parcelable {
         dest.writeString(accountRegion);
         dest.writeString(proxyDomain);
         dest.writeString(spikyProxyDomain);
+        dest.writeString(customHandshakeDomain);
         dest.writeByte((byte) (analyticsOnly ? 0x01 : 0x00));
         dest.writeByte((byte) (isDefaultInstance ? 0x01 : 0x00));
         dest.writeByte((byte) (useGoogleAdId ? 0x01 : 0x00));
@@ -474,6 +490,7 @@ public class CleverTapInstanceConfig implements Parcelable {
             configJsonObject.put(Constants.KEY_ACCOUNT_REGION, getAccountRegion());
             configJsonObject.put(Constants.KEY_PROXY_DOMAIN, getProxyDomain());
             configJsonObject.put(Constants.KEY_SPIKY_PROXY_DOMAIN, getSpikyProxyDomain());
+            configJsonObject.put(Constants.KEY_CUSTOM_HANDSHAKE_DOMAIN, getCustomHandshakeDomain());
             configJsonObject.put(Constants.KEY_FCM_SENDER_ID, getFcmSenderId());
             configJsonObject.put(Constants.KEY_ANALYTICS_ONLY, isAnalyticsOnly());
             configJsonObject.put(Constants.KEY_DEFAULT_INSTANCE, isDefaultInstance());